+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 6.3.2
+
+- Extend `AbstractFeatureInfoView`
+  - Add new option `mergeParentAttributes`, which will merge attributes from parent features, if attribute `__PARENT_FEATURE` is set on the feature. Child attributes will overwrite parent attributes in case of identical keys.
+  - Add new option `removeNoDataAttributes` to `AbstractFeatureInfoView`, which filters all attributes with no data values (undefined).
+- Add recursive attribute fetching for child features having a `__PARENT_FEATURE`. Found parent feature attributes are merged.
+
 # 6.3.1
 
 ### Fixes